+++ SuperMarioBros-3/Camera.cpp
@@ -5,7 +5,7 @@ Camera::Camera(CMario* player)
 	this->player = player;
 }
 
-void Camera::Update(DWORD dt)
+void Camera::Update(ULONGLONG dt)
 {
 	x += vx * dt;
 }
