+++ SuperMarioBros-3/Effect.cpp
@@ -11,7 +11,7 @@ void CEffect::GetBoundingBox(float& l, float& t, float& r, float& b)
 	l = t = r = b = 0;
 }
 
-void CEffect::Update(DWORD dt, vector<LPGAMEOBJECT>* objects)
+void CEffect::Update(ULONGLONG dt, vector<LPGAMEOBJECT>* objects)
 {
 	if (GetTickCount64() - startEffect > effectTime)
 		isFinishedUsing = true;
