+++ SuperMarioBros-3/Effect.h
@@ -9,7 +9,7 @@ protected:
 public:
 	CEffect();
 	virtual void GetBoundingBox(float& l, float& t, float& r, float& b);
-	virtual void Update(DWORD dt, vector<LPGAMEOBJECT>* objects);
+	virtual void Update(ULONGLONG dt, vector<LPGAMEOBJECT>* objects);
 	virtual void Render() = 0;
 	~CEffect();
 };
